refactor: rename integrations directory to interceptors
Move dio_interceptor and its test from src/integrations to src/interceptors
and update the export/import references accordingly. The new name better
reflects that the directory holds request interceptors.
