+++ lib/flutter_inspector_kit.dart
@@ -1,5 +1,5 @@
 export 'src/core/flutter_inspector_impl.dart';
-export 'src/integrations/dio_interceptor.dart';
+export 'src/interceptors/dio_interceptor.dart';
 export 'src/models/database_browser_source.dart';
 export 'src/models/database_operation.dart';
 export 'src/models/log_level.dart';
